Minor doc changes
