+++ rubrik_polaris/lib/common/graphql/mutation_instances_restore_ec2.graphql
@@ -1,4 +1,4 @@
-mutation RubrikPolarisSDKRequest($snapshot_id: String!, $should_power_on: Boolean!, $should_restore_tags: Boolean!) {
+mutation RubrikPolarisSDKRequest($snapshot_id: String!, $should_power_on: Boolean = true, $should_restore_tags: Boolean = true) {
   restoreAwsNativeEc2InstanceSnapshot(snapshotId: $snapshot_id, shouldPowerOn: $should_power_on, shouldRestoreTags: $should_restore_tags) {
     taskchainUuid
   }
