+++ rubrik_polaris/lib/compute.py
@@ -115,3 +115,75 @@ def get_instances_gce(self):
         return self._dump_nodes(_request, _query_name)
     except Exception as e:
         print(e)
+
+
+def _submit_instance_restore(self, snapshot_id, **kwargs):
+    """Submits a Restore of a compute instance
+
+    Arguments:
+        query_name [string] -- Backend query name for operation
+        snapshot_id [string] -- Snapshot ID to be restored
+        should_power_on bool -- Defaults to False
+        should_restore_tags bool -- Defaults to False
+        wait bool -- Return once complete Defaults to False
+    """
+
+    should_power_on = False
+    if kwargs and 'should_power_on' in kwargs and kwargs['should_power_on']:
+        should_power_on = True
+    should_restore_tags = False
+    if kwargs and 'should_restore_tags' in kwargs and kwargs['should_restore_tags']:
+        should_power_on = True
+    try:
+        _mutation_name = kwargs['mutation']
+        _variables = {
+            "snapshot_id": snapshot_id,
+            "should_power_on": kwargs['should_power_on'],
+            "should_restore_tags": kwargs['should_restore_tags']
+        }
+        if _mutation_name not in self._graphql_query:
+            raise Exception("Mutation not found : {}".format(_mutation_name))
+        _request = self._query(None, self._graphql_query[_mutation_name], _variables)
+        if 'errors' in _request and _request['errors']:
+            return  {'errors': _request['errors'][0]['message']}
+        _result = self._dump_nodes(_request, _mutation_name)
+        _results = []
+        if 'wait' in kwargs:
+            _results = self._monitor_task(_result)
+        return _results
+        #todo: find a better way to report errors per uuid
+    except Exception as e:
+        print(e)
+
+def submit_restore_ec2(self, snapshot_id, **kwargs):
+    """Submits a Restore of an EC2 instance
+
+    Arguments:
+        snapshot_id [string] -- Snapshot ID to be restored
+        should_power_on bool -- Defaults to False
+        should_restore_tags bool -- Defaults to False
+        wait bool -- Return once complete Defaults to False
+    """
+    return self._submit_instance_restore(snapshot_id, mutation = "instances_restore_ec2", **kwargs)
+
+def submit_restore_azure(self, snapshot_id, **kwargs):
+    """Submits a Restore of an Azure instance
+
+    Arguments:
+        snapshot_id [string] -- Snapshot ID to be restored
+        should_power_on bool -- Defaults to False
+        should_restore_tags bool -- Defaults to False
+        wait bool -- Return once complete Defaults to False
+    """
+    return self._submit_instance_restore(snapshot_id, mutation = "instances_restore_azure", **kwargs)
+
+def submit_restore_gce(self, snapshot_id, **kwargs):
+    """Submits a Restore of a GCE instance
+
+    Arguments:
+        snapshot_id [string] -- Snapshot ID to be restored
+        should_power_on bool -- Defaults to False
+        should_restore_tags bool -- Defaults to False
+        wait bool -- Return once complete Defaults to False
+    """
+    return self._submit_instance_restore(snapshot_id, mutation = "instances_restore_gce", **kwargs)
