+++ rubrik_polaris/rubrik_polaris.py
@@ -28,14 +28,16 @@ import pprint
 class PolarisClient:
     # Public
     from .lib.common.polaris import get_sla_domains, submit_on_demand, submit_assign_sla, get_task_status
-    from .lib.common.polaris import get_snapshots, submit_restore_ec2
+    from .lib.common.polaris import get_snapshots
     from .lib.compute import get_instances_azure, get_instances_ec2, get_instances_gce
+    from .lib.compute import submit_restore_ec2, submit_restore_azure, submit_restore_gce
     from .lib.accounts import get_accounts_aws, get_accounts_azure, get_accounts_gcp, delete_account_aws
     from .lib.accounts import add_account_aws, get_accounts_aws_detail, get_account_aws_native_id
     from .lib.compute import get_object_ids_azure, get_object_ids_ec2, get_object_ids_gce
 
     # Private
     from .lib.common.connection import _query, _get_access_token
+    from .lib.compute import _submit_instance_restore
     from .lib.common.monitor import _monitor_job, _monitor_threader, _monitor_task
     from .lib.common.graphql import _dump_nodes, _get_query_names_from_graphql_query
     from .lib.accounts import _invoke_account_delete_aws, _invoke_aws_stack, _commit_account_delete_aws
@@ -91,7 +93,7 @@ class PolarisClient:
         }
 
         # Get graphql content
-        (self._graphql_query, self._graphql_mutation, self._graphql_file_type_map) = _build_graphql_maps(self)
+        (self._graphql_query, self._graphql_file_type_map) = _build_graphql_maps(self)
 
     def _log(self, log_message):
         """Create properly formatted debug log messages.
