+++ sample/polaris_client.py
@@ -17,7 +17,6 @@ parser.add_argument('-r', '--root', dest='root_domain', help="Polaris Root Domai
 parser.add_argument('--insecure', help='Deactivate SSL Verification', action="store_true")
 
 args = parser.parse_args()
-
 rubrik = rubrik_polaris.PolarisClient(args.domain, args.username, args.password, root_domain=args.root_domain,
                                       insecure=args.insecure)
 
@@ -28,16 +27,26 @@ rubrik = rubrik_polaris.PolarisClient(args.domain, args.username, args.password,
 # pp.pprint(rubrik.delete_account_aws())
 
 ### Run ODS for machines in a region using Gold retention, monitor to complete via threads
-pp.pprint(rubrik.submit_on_demand(rubrik.get_object_ids_gce(region="us-west1"), rubrik.get_sla_domains("Gold"), wait = True))
+# pp.pprint(rubrik.submit_on_demand(rubrik.get_object_ids_gce(region="us-west1"), rubrik.get_sla_domains("Gold"), wait = True))
+
+### Returns all objectIDs matching arbitrary available inputs. ec2 tags have special treatment
+# pp.pprint(rubrik.get_object_ids_ec2(tags = {"Name": "gurlingwinjb"}))
+# pp.pprint(rubrik.get_object_ids_azure(region = "EastUS2"))
+# pp.pprint(rubrik.get_object_ids_gce(region = "us-west1"))
 
 ### Get snapshot ids for snappables
 # snappables = rubrik.get_object_ids_ec2(tags = {"Name": "gurlingwinjb"})
+# snappables = rubrik.get_object_ids_azure(name = "tpm1-lin1")
+# snappables = rubrik.get_object_ids_gce(nativeName = "ubuntu-fdse-shared-1")
 # for snappable in snappables:
-#     pp.pprint(rubrik.get_snapshots(snappable, recovery_point='2020-09-19 04:20')) # can include anything up to this. 2020 is ok, 2020-09, 2020-09-19, ...
+#     snapshot_id = (rubrik.get_snapshots(snappable, recovery_point='2020-09-19 04:20')) # can include anything up to this. 2020 is ok, 2020-09, 2020-09-19, ...
 #     snapshot_id = rubrik.get_snapshots(snappable, recovery_point='latest')['id']
-#     pp.pprint(rubrik.get_snapshots(snappable))
+#     pp.pprint(rubrik.get_snapshots(snappable)) # Get all snapshots
+
 ### Submit Restore for above Snapshot (EC2)
 #     result = rubrik.submit_restore_ec2(snapshot_id, wait=True, should_power_on=True, should_restore_tags=True)
+#     result = rubrik.submit_restore_azure(snapshot_id, wait=True, should_power_on=True, should_restore_tags=True)
+#     result = rubrik.submit_restore_gce(snapshot_id, wait=True, should_power_on=True, should_restore_tags=True)
 #     pp.pprint(result)
 
 ### Search for a set of objects and get their details
@@ -63,10 +72,5 @@ pp.pprint(rubrik.submit_on_demand(rubrik.get_object_ids_gce(region="us-west1"), 
 
 ### Query objects, set sla_domain
 # pp.pprint(rubrik.submit_assign_sla( rubrik.get_object_ids_ec2(region = "US_WEST_2"), rubrik.get_sla_domains("Gold")))
-
-### Returns all objectIDs matching arbitrary available inputs. ec2 tags have special treatment
-# pp.pprint(rubrik.get_object_ids_ec2(tags = {"Name": "gurlingwinjb"}))
-# pp.pprint(rubrik.get_object_ids_azure(region = "EastUS2"))
-# pp.pprint(rubrik.get_object_ids_gce(region = "us-west1"))
